refactor: Use re-export
