+++ src/App/Controllers/CommandController.ts
@@ -1,6 +1,6 @@
 import Controller from "../../Core/Controller/Controller";
 import CommandHandler from "../../Core/Handlers/CommandHandler";
-import PingCommand from "../Handlers/Command/PingCommand";
+import { PingCommand } from "../Handlers/Command/";
 
 export default class CommandController extends Controller<CommandHandler> {
     protected RegisterHandlers(): void {
