+++ src/App/Handlers/Command/index.ts
@@ -1,0 +1,1 @@
+export { default as PingCommand } from "./PingCommand";
