+++ src/App/Handlers/Message/index.ts
@@ -1,0 +1,1 @@
+export { default as HelloMessage } from "./HelloMessage";
